requirements.txt
Django==1.11.5
django-environ==0.4.4
djangorestframework==3.6.4
psycopg2==2.7.3.1
pytz==2017.2
six==1.10.0
